Problem: Telegram doesn't have voice calls
That kind of sucks

Solution: try pulling it from nixpkgs-unstable

I am pretty sure what I am doing with fetchTarball is a horrible
hack and I will pay dearly for this. But I'll figure out the right
way to do this one day.

For now, this lets me to have calls on Telegram.