Problem: IDEA can't remember recent project
This happens because it tries to save to
sub-directories of its config, which are linked
into the nix store and, therefore, read-only.

Solution: provision IDEA configuration as symlink
to regular files only, by traversing the config
directory recursively.

Note that this method is still faulty as it won't
work with files that contain spaces (or Unicode
characters). Such files are currenly removed in
my local copy to enable config provisioning, but
there should be a better solution to just dropping
such files.